requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.5
cryptography==46.0.2
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.12.0
django-celery-beat==2.8.1
django-js-asset==3.1.2
django-mptt==0.17.0
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
flower==2.0.1
humanize==4.13.0
idna==3.10
inflection==0.5.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
